# Conflicts:
#	erms-project1/src/main/java/com/revature/erms_project1/controllers/AccountController.java
#	erms-project1/src/main/java/com/revature/erms_project1/entities/Account.java

+++ erms-project1/src/main/java/com/revature/erms_project1/controllers/UserController.java
@@ -39,7 +39,7 @@ public class UserController {
     }
 
     @GetMapping("/users/{userId}")
-    public ResponseEntity<User> getUser(@PathVariable("userId") Long personId) {
+    public ResponseEntity<User> getUser(@PathVariable("userId") Integer personId) {
         try{
             User user = this.userService.getUserById(personId);
             return new ResponseEntity<>(user, HttpStatus.OK);
