+++ erms-project1/src/main/java/com/revature/erms_project1/repositories/UserRepository.java
@@ -3,6 +3,10 @@ package com.revature.erms_project1.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.revature.erms_project1.entities.User;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+//public interface UserRepository extends JpaRepository<User, Long> {
+//    User findByUsername(String username);
+//}
+
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByUsername(String username);
 }
